split into pk and log table source tests
